+++ src/include/cpu_prefetch.h
@@ -58,7 +58,7 @@ STATIC_INLINE void do_cycles_ce020_internal(int clocks)
 		return;
 	}
 	int cycs = clocks * cpucycleunit;
-	int diff = regs.ce020endcycle - regs.ce020startcycle;
+	int diff = (int)(regs.ce020endcycle - regs.ce020startcycle);
 	if (diff > 0) {
 		if (diff >= cycs) {
 			regs.ce020startcycle += cycs;
